Move RecoverDatabaseFile and RecoverKeysOnlyFilter into salvage.{cpp/h}
