+++ src/wallet/db.h
@@ -389,8 +389,6 @@ public:
     bool static Rewrite(BerkeleyDatabase& database, const char* pszSkip = nullptr);
 };
 
-bool RecoverDatabaseFile(const fs::path& file_path, void *callbackDataIn, bool (*recoverKVcallback)(void* callbackData, CDataStream ssKey, CDataStream ssValue), std::string& out_backup_filename);
-
 std::string BerkeleyDatabaseVersion();
 
 #endif // BITCOIN_WALLET_DB_H
